pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chronos-x"
version = "0.1.0"
description = "Crypto trading system with Mamba-based ML models"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Chronos-X Team" }
]

dependencies = [
    # Configuration Management
    "hydra-core>=1.3",
    "omegaconf>=2.3",

    # Database
    "clickhouse-driver>=0.2",
    "clickhouse-connect>=0.7",

    # Data Exchange APIs
    "ccxt>=4.0",
    "websockets>=12.0",

    # Data Processing
    "polars>=0.20",
    "numpy>=1.26",
    "numba>=0.59",
    "pandas>=2.0",
    "scipy>=1.12",

    # Machine Learning - Core
    "torch>=2.2",
    "scikit-learn>=1.4",

    # Gradient Boosting
    "xgboost>=2.0",
    "lightgbm>=4.0",

    # Optimization
    "optuna>=3.5",

    # Backtesting
    "vectorbt>=0.26",

    # Risk Management
    "riskfolio-lib>=6.0",

    # MLOps
    "mlflow>=2.10",

    # Async & Web
    "aiohttp>=3.9",
    "httpx>=0.27",

    # Utilities
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "rich>=13.0",
    "typer>=0.9",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
]

[project.optional-dependencies]
mamba = [
    "mamba-ssm>=2.0",
    "causal-conv1d>=1.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[project.scripts]
chronos-ingest = "chronos_x.scripts.ingest_data:main"
chronos-train = "chronos_x.scripts.train_model:main"
chronos-backtest = "chronos_x.scripts.run_backtest:main"

[tool.hatch.build.targets.wheel]
packages = ["chronos_x"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
